added portability for MSVC

+++ GLGE/include/Graphic/Backend/Builtin/Graphics/OpenGL/MeshPool.h
@@ -259,7 +259,7 @@ namespace GLGE::Graphic::Backend::Graphic::OpenGL {
         /**
          * @brief store the size of the vertex buffer in bytes
          */
-        u64 m_vboSize = 0;
+        size_t m_vboSize = 0;
         /**
          * @brief store the actual index buffer
          */
@@ -277,7 +277,7 @@ namespace GLGE::Graphic::Backend::Graphic::OpenGL {
         /**
          * @brief store the size of the index buffer in bytes
          */
-        u64 m_iboSize = 0;
+        size_t m_iboSize = 0;
 
     };
